+++ assemblies/core/lib/pom.xml
@@ -53,7 +53,6 @@
         <databricks.version>4.0.0</databricks.version>
         <paho.version>1.2.0</paho.version>
         <bahir.version>2.1.1</bahir.version>
-        <xerces.version>2.12.2</xerces.version>
 
         <!-- Spring -->
         <springframework4.version>4.3.18</springframework4.version>
@@ -100,12 +99,6 @@
                     <artifactId>*</artifactId>
                 </exclusion>
             </exclusions>
-        </dependency>
-
-        <!-- Third-party JDBC dependencies -->
-        <dependency>
-            <groupId>jaxen</groupId>
-            <artifactId>jaxen</artifactId>
         </dependency>
 
         <!-- Other third-party dependencies -->
@@ -290,24 +283,6 @@
         <dependency>
             <groupId>xerces</groupId>
             <artifactId>xercesImpl</artifactId>
-            <version>${xerces.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>*</groupId>
-                    <artifactId>*</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>xml-apis</groupId>
-            <artifactId>xml-apis</artifactId>
-            <version>1.4.01</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>*</groupId>
-                    <artifactId>*</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
 
     </dependencies>
