+++ assemblies/lib/pom.xml
@@ -102,13 +102,12 @@
                 </exclusion>
             </exclusions>
         </dependency>
-        <!-- Third-party JDBC dependencies -->
-        <dependency>
-            <groupId>jaxen</groupId>
-            <artifactId>jaxen</artifactId>
-        </dependency>
 
         <!-- Other third-party dependencies -->
+        <dependency>
+            <groupId>xerces</groupId>
+            <artifactId>xercesImpl</artifactId>
+        </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.jaxrs</groupId>
             <artifactId>jackson-jaxrs-json-provider</artifactId>
