+++ assemblies/plugins/transforms/xml/pom.xml
@@ -38,7 +38,8 @@
 
     <properties>
         <saxon.version>8.7</saxon.version>
-        <jaxen.version>1.1.6</jaxen.version>
+        <dom4j.version>2.1.3</dom4j.version>
+        <jaxen.version>1.2.0</jaxen.version>   
     </properties>
 
     <dependencies>
@@ -52,6 +53,11 @@
             <artifactId>saxon-dom</artifactId>
             <version>${saxon.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.dom4j</groupId>
+            <artifactId>dom4j</artifactId>
+            <version>${dom4j.version}</version>
+        </dependency>        
         <dependency>
             <groupId>jaxen</groupId>
             <artifactId>jaxen</artifactId>
