+++ core/pom.xml
@@ -61,7 +61,6 @@
         <json-simple.version>1.1.1</json-simple.version>
         <gson.version>2.8.5</gson.version>
         <jandex.version>2.2.2.Final</jandex.version>
-        <xerces.version>2.12.2</xerces.version>
     </properties>
 
     <dependencies>
@@ -414,10 +413,6 @@
             <artifactId>commons-compress</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.dom4j</groupId>
-            <artifactId>dom4j</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.jboss</groupId>
             <artifactId>jandex</artifactId>
             <version>${jandex.version}</version>
@@ -485,7 +480,13 @@
             <artifactId>woodstox-core-asl</artifactId>
             <version>${woodstox-core-asl.version}</version>
             <scope>test</scope>
-        </dependency>
+            <exclusions>
+               <exclusion>
+                 <groupId>javax.xml.stream</groupId>
+                 <artifactId>stax-api</artifactId>
+               </exclusion>
+             </exclusions>
+        </dependency>        
         <dependency>
             <groupId>org.hamcrest</groupId>
             <artifactId>java-hamcrest</artifactId>
@@ -516,12 +517,10 @@
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-core</artifactId>
-            <version>${jackson.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
