+++ plugins/transforms/excelwriter/pom.xml
@@ -35,7 +35,6 @@
 
     <properties>
         <poi.version>5.2.1</poi.version>
-        <xerces.version>2.12.2</xerces.version>
     </properties>
 
     <dependencies>
@@ -52,7 +51,6 @@
         <dependency>
             <groupId>xerces</groupId>
             <artifactId>xercesImpl</artifactId>
-            <version>${xerces.version}</version>
         </dependency>
         <dependency>
             <groupId>org.mockito</groupId>
