+++ plugins/transforms/xml/src/main/java/org/apache/hop/pipeline/transforms/xml/getxmldata/XmlInputFieldsImportProgressDialog.java
@@ -23,8 +23,8 @@ import org.apache.hop.core.RowMetaAndData;
 import org.apache.hop.core.row.IValueMeta;
 import org.apache.hop.core.util.Utils;
 import org.apache.hop.core.vfs.HopVfs;
-import org.apache.hop.core.xml.XmlParserFactoryProducer;
 import org.apache.hop.i18n.BaseMessages;
+import org.apache.hop.pipeline.transforms.xml.Dom4JUtil;
 import org.apache.hop.ui.core.dialog.ErrorDialog;
 import org.apache.hop.ui.core.dialog.ProgressMonitorDialog;
 import org.dom4j.Attribute;
@@ -155,7 +155,7 @@ public class XmlInputFieldsImportProgressDialog {
             PKG, "GetXMLDateLoopNodesImportProgressDialog.Task.ScanningFile", filename),
         1);
 
-    SAXReader reader = XmlParserFactoryProducer.getSAXReader(null);
+    SAXReader reader = Dom4JUtil.getSAXReader(null);
     monitor.worked(1);
     if (monitor.isCanceled()) {
       return null;
