+++ pom.xml
@@ -168,8 +168,6 @@
 
         <codehaus-jackson.version>1.9.13</codehaus-jackson.version>
         <jackson.version>2.13.0</jackson.version>
-        <dom4j.version>2.1.3</dom4j.version>
-        <jaxen.version>1.1.6</jaxen.version>
         <commons-compress.version>1.21</commons-compress.version>
         <commons-fileupload.version>1.4</commons-fileupload.version>
         <aws-java-sdk.version>1.11.974</aws-java-sdk.version>
@@ -178,7 +176,6 @@
         <spring-security.version>4.1.5.RELEASE</spring-security.version>
 
         <xerces.version>2.12.2</xerces.version>
-        <xml-apis.version>1.4.01</xml-apis.version>
 
         <!-- jdk version -->
         <source.jdk.version>11</source.jdk.version>
@@ -420,17 +417,6 @@
                 <artifactId>jackson-jr-objects</artifactId>
                 <version>${jackson.version}</version>
             </dependency>
-
-            <dependency>
-                <groupId>org.dom4j</groupId>
-                <artifactId>dom4j</artifactId>
-                <version>${dom4j.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>jaxen</groupId>
-                <artifactId>jaxen</artifactId>
-                <version>${jaxen.version}</version>
-            </dependency>
             <dependency>
                 <groupId>commons-fileupload</groupId>
                 <artifactId>commons-fileupload</artifactId>
@@ -470,12 +456,13 @@
                 <groupId>xerces</groupId>
                 <artifactId>xercesImpl</artifactId>
                 <version>${xerces.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>xml-apis</groupId>
-                <artifactId>xml-apis</artifactId>
-                <version>${xml-apis.version}</version>
-            </dependency>
+                <exclusions>
+                  <exclusion>
+                    <groupId>*</groupId>
+                    <artifactId>*</artifactId>
+                  </exclusion>
+              </exclusions>
+            </dependency>            
             <dependency>
                 <groupId>org.apache.commons</groupId>
                 <artifactId>commons-compress</artifactId>
